"Properly" handle locale detection with subrequests
The previous `$app->before()` filter only worked for the master request.
Moved the code to an event subscriber, which gets executed for every
request, no matter the type.

We also can't rely on the `$app['locale']` value anymore.
See GitHub issue #384 for Silex about that.

The goal of the event listener is to get the {locale} attribute
from the request and set it as the request's locale (if it's valid).

The listener also sets the locale on the translator, which will fail if
we do subrequests with different locales. If we start with locale "bg"
and do a subrequests for "en", "en" will be set on the translator even
after we get back from the subrequest. Nothing restores it back to "bg".
This is a known problem.

Most things try and should use `$request->getLocale()` now to get an
accurate current locale.